+++ app/cdap/components/GridTable/services.ts
@@ -1,0 +1,42 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import { directiveRequestBodyCreator } from 'components/DataPrep/helper';
+import DataPrepStore from 'components/DataPrep/store';
+
+export const getAPIRequestPayload = (params, newDirective, action) => {
+  const { dataprep } = DataPrepStore.getState();
+  const { workspaceId, workspaceUri, directives, insights } = dataprep;
+  let gridParams = {};
+  const updatedDirectives = directives.concat(newDirective);
+  const requestBody = directiveRequestBodyCreator(updatedDirectives);
+  requestBody.insights = insights;
+  const workspaceInfo = {
+    properties: insights,
+  };
+  gridParams = {
+    directives: updatedDirectives,
+    workspaceId,
+    workspaceUri,
+    workspaceInfo,
+    insights,
+  };
+  const payload = {
+    context: params.namespace,
+    workspaceId: params.wid,
+  };
+  return { payload, requestBody, gridParams };
+};
